Move /products/qc into a products router so that it doesn't have security restrictions
